changing doubles to floats

+++ C/src/CHeterodyning.c
@@ -1,13 +1,13 @@
 #include "CHeterodyning.h"
 
-extern double data [SAMPLE_COUNT];
-extern double carrier[SAMPLE_COUNT];
+extern float data [SAMPLE_COUNT];
+extern float carrier[SAMPLE_COUNT];
 
-double result [SAMPLE_COUNT];
+float result [SAMPLE_COUNT];
 
 int main(int argc, char**argv){
     printf("Running Unthreaded Test\n");
-    printf("Precision sizeof %ld\n", sizeof(double));
+    printf("Precision sizeof %ld\n", sizeof(float));
     
 
     printf("Total amount of samples: %ld\n", sizeof(data) / sizeof(data[0]));
